Move ApiRequest to bot.irc package
